requirements.txt
Django==3.2
django-ckeditor==6.1.0
requests==2.25.1



